checkJWK :: (MonadError e m, AsError e) => JWK -> m ()
Add the 'checkJWK' function for sanity checking a JWK to make sure
it is not totally unusable.  This work is motivated by the problem
where loading a small key means that signature verification fails
with 'InvalidSignature', obscuring the fact that the verification
key is too small to be used.

The fact that the validation function may attempt to validate a
single signature against several keys makes it difficult to convey
the error for a single signature to the caller.  Instead, library
users are encouraged to explicitly apply checkJWK to all JWKs that
may be used for verification, and fail, warn, or omit offending
keys.

Fixes: https://github.com/frasertweedale/hs-jose/issues/46